Pulled middleware stuff out into a driver
